Getting joystick info (#77); Safety Limit to thruster mapper
* Nodes for getting joystick info

* fixed folder structure and precommit errors

* modified joystick bindings

* decreased sim step size and decreased sub mass

* scaled up the joy wrench output

* added safety limit to thruster_manager

* refactored keyboard control code

---------
