+++ src/subjugator/gnc/subjugator_thruster_manager/CMakeLists.txt
@@ -36,4 +36,9 @@ install(DIRECTORY
   DESTINATION share/${PROJECT_NAME}
 )
 
+install(DIRECTORY
+  include/
+  DESTINATION include/${PROJECT_NAME}
+)
+
 ament_package()
