+++ src/subjugator/gnc/subjugator_thruster_manager/config/thruster_manager.yaml
@@ -10,3 +10,4 @@
       rx: [0.0,           0.0,            0.0,            0.0,          -0.14922, 0.14922,  -0.14922, 0.14922 ] # roll
       ry: [0.0,           0.0,            0.0,            0.0,          -0.33337, -0.33337, 0.127,    0.127   ] # pitch
       rz: [0.33106057,    -0.33106057,    0.277971191,    -0.277971191, 0.0,      0.0,      0.0,      0.0     ] # yaw
+    safety_limit: 200.0
