+++ src/subjugator/simulation/subjugator_gazebo/worlds/robosub_2024.world
@@ -13,7 +13,7 @@
             <!-- Physics Settings -->
             <physics name='default_physics' default='0' type='ode'>
                 <real_time_update_rate>200</real_time_update_rate>
-                <max_step_size>0.005</max_step_size>
+                <max_step_size>0.003</max_step_size>
                 <real_time_factor>1.0</real_time_factor>
             </physics>
